# Conflicts:
#	src/java_/game/controller/GameService.java

+++ src/java_/game/tile/FloorTile.java
@@ -101,7 +101,7 @@ public class FloorTile extends Tile {
                 '}';
     }
 
-    public int getPathsBits() {
+    public byte getPathsBits() {
         return pathsBits;
     }
